lab4/deploy/requirements.txt
pycryptodome==3.17
sympy==1.11.1
